The watchdog python program
This program runs on a RPi (the home automation server typically) and listens for TCP connections from various computers or applications.  When a connection is made, the program responds with a serial number, then sends an MQTT message to the HA server.  The HA server uses these messages to display the status of the computer/application in Home Assistant